# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация по API Arenatickets для партнеров",
...
        "contact": {
            "email": "hello@spaceduck.kz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://testapi.arenatickets.kz",
            "description": "Develop API Server"
        },
        {
            "url": "https://api.arenatickets.kz",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/partner/shows/{id}": {
            "get": {
                "summary": "Детали по выбранному событию",
                "description": "Выдает количество доступных для продажи билетов по каждому сектору / ценовой группе",
                "operationId": "a1f0f5b95892f3092c3516ccc576f3f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id события",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tickets": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SectionInfo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/partner/shows/{id}/section/{sectionId}": {
            "get": {
                "summary": "Билеты по выбранному сектору",
                "description": "Выдает список доступных для продажи билетов по сектору – билеты могут быть по местам (параметры row и seat имеют значения) или входные для сектора со свободной рассадкой (параметры row и seat равны null)",
                "operationId": "ae58e2f0a62792d51b270b48a4952172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id события",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sectionId",
                        "in": "path",
                        "description": "Id сектора или ценовой группы для событий по ценовым группам (если не передается, то выдает список всех секторов или ценовых категорий)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tickets": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ticket"
                                            }
                                        },
                                        "seats": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Seat"
                                            }
                                        },
                                        "prices": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": {
                                            "description": "pricegroups для ценовых групп, enter - для сектора с входными без мест (например, фан зона), seats - для сектора по местам, ",
                                            "type": "string",
                                            "example": "seats"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/partner/shows/{id}/order": {
            "post": {
                "summary": "Инициирование заказа",
                "description": "Генерация заказа на этапе, когда клиент выбрал билеты и перешел к процедуре оплаты, но еще не оплатил. Бронь на выбранные билеты держится 30 минут. В случае успешного запроса ответом приходит id созданного заказа. Если оплаты не происходит, через указанный интервал времени заказ удаляется, и билеты высвобождаются для продажи",
                "operationId": "316743aca210d29847fb49b5b9ea9ce0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id события",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cart": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/CartItem"
                                        }
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Ермек",
                                        "description": "Имя пользователя"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+77011234567",
                                        "description": "Телефон пользователя"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "email@example.com",
                                        "description": "Email пользователя"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/partner/order/{id}": {
            "get": {
                "summary": "Получение статуса заказа",
                "operationId": "d2a3f3808be8c4e225e4f6fa14f9c1a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/partner/order/{id}/confirm": {
            "post": {
                "summary": "Подтверждение оплаты заказа",
                "description": "Вызывается, когда оплата получена, и билеты необходимо выкупить",
                "operationId": "ee4e716cbb6176a8d3e2a5d5af04e773",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/partner/order/{id}/cancel": {
            "delete": {
                "summary": "Отмена заказа",
                "description": "Применимо только к неоплаченному заказу",
                "operationId": "20113e81ecd82e9ac5c97c56dcf5e104",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/partner/order/{id}/refund": {
            "delete": {
                "summary": "Возврат заказа",
                "description": "Применимо только к оплаченному заказу. Заказ попадает в статистику возвратов",
                "operationId": "afb3771520caac01f1b4e30c1f7f2c07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SectionInfo": {
                "title": "Общая информация по сектору / ценовой группе",
                "properties": {
                    "amount": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "Бывает два типа: sections или pricegroups, по секторам или по ценовым группам",
                        "type": "string",
                        "example": "sections"
                    },
                    "pricegroup": {
                        "$ref": "#/components/schemas/Pricegroup"
                    },
                    "section": {
                        "$ref": "#/components/schemas/Section"
                    }
                },
                "type": "object"
            },
            "Pricegroup": {
                "title": "Ценовая группа",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "$ref": "#/components/schemas/Translation"
                    },
                    "price": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
            "Seat": {
                "title": "Место",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "x": {
                        "description": "x координата",
                        "type": "integer",
                        "example": 100
                    },
                    "y": {
                        "description": "y координата",
                        "type": "integer",
                        "example": 100
                    },
                    "row": {
                        "description": "номер / название ряда",
                        "type": "string",
                        "example": "1"
                    },
                    "seat": {
                        "description": "номер / название места",
                        "type": "string",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "title": "Билет",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "row": {
                        "description": "номер / название ряда",
                        "type": "string",
                        "example": "1"
                    },
                    "seat": {
                        "description": "номер / название места",
                        "type": "string",
                        "example": "1"
                    },
                    "seat_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "integer",
                        "example": 100
                    },
                    "blocked": {
                        "description": "Блокирован ли билет (те, забронирован, но не выкуплен)",
                        "type": "boolean",
                        "example": false
                    },
                    "sold": {
                        "description": "Продан ли билет",
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "CartItem": {
                "title": "Билет корзины",
                "properties": {
                    "ticket_id": {
                        "description": "Id билета (именно билета, не места). Параметр нужен только при рассадке по местам (type = seats из запроса по сектору)",
                        "type": "integer",
                        "example": 1
                    },
                    "section_id": {
                        "description": "Id сектора. Параметр нужен только при входных по сектору без места (type = enter из запроса по сектору)",
                        "type": "integer",
                        "example": 1
                    },
                    "pricegroup_id": {
                        "description": "Id ценовой группы. Параметр нужен только при покупке билетов по ценовым группам (тип события - pricegroups)",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "title": "Ошибка",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Неверный формат данных"
                    }
                },
                "type": "object"
            },
            "NotFoundError": {
                "title": "Ошибка",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Запись не найдена"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "description": "Токен выдается менеджером Arenatickets",
                "name": "X-PARTNER-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "api_token": []
        }
    ]
}